Move `node_base` into `p2p`
