+++ nucypher_async/p2p/__init__.py
@@ -1,3 +1,31 @@
-from .node_info import NodeClient
+from .errors import GenericPeerError, InactivePolicy, InvalidMessage, PeerError
+from .fleet_sensor import FleetSensorSnapshot, NodeEntry
+from .keys import Contact, PeerPrivateKey, PeerPublicKey, SecureContact, get_alternative_contact
+from .learner import Learner
+from .node_client import NodeClient
+from .node_info import NodeInfo
+from .operator import Operator
+from .routes import NodeRoutes
+from .verification import PeerVerificationError, VerifiedNodeInfo, verify_staking_local
 
-__all__ = ["NodeClient"]
+__all__ = [
+    "Contact",
+    "FleetSensorSnapshot",
+    "GenericPeerError",
+    "InactivePolicy",
+    "InvalidMessage",
+    "Learner",
+    "NodeClient",
+    "NodeEntry",
+    "NodeInfo",
+    "NodeRoutes",
+    "Operator",
+    "PeerError",
+    "PeerPrivateKey",
+    "PeerPublicKey",
+    "PeerVerificationError",
+    "SecureContact",
+    "VerifiedNodeInfo",
+    "get_alternative_contact",
+    "verify_staking_local",
+]
