+++ tests/conftest.py
@@ -7,7 +7,6 @@ import pytest
 import trio
 
 from nucypher_async.characters.cbd import Decryptor
-from nucypher_async.characters.node import Operator
 from nucypher_async.characters.pre import Reencryptor
 from nucypher_async.domain import Domain
 from nucypher_async.drivers.identity import AmountT, IdentityAccount, IdentityAddress
@@ -24,7 +23,7 @@ from nucypher_async.mocks import (
     MockPREClient,
 )
 from nucypher_async.node import HTTPServerConfig, NodeServer, NodeServerConfig, SSLConfig
-from nucypher_async.node_base import Contact
+from nucypher_async.p2p import Contact, Operator
 from nucypher_async.proxy import ProxyServer, ProxyServerConfig
 from nucypher_async.storage import InMemoryStorage
 from nucypher_async.utils import logging
